app/pyproject.toml
[project]
name = "mailfilter"
version = "0.1.0"
description = "MailFilter — stop physical junk mail"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic-settings",
    "python-dotenv",
    "anthropic",
    "google-api-python-client",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "cryptography",
    "arq",
    "supabase",
    "jinja2",
    "httpx",
    "resend",
    "python-multipart",
    "itsdangerous",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
